Mouse function
I added the ability to play with keyboard and mouse but it's pretty wonky right now. If we want it to be better I can fix it, but for right now I made it so that it would be possible to play with keyboard and mouse. WASD is to move, use the mouse to look around, left click to attack.