Level One Is Done
Notable Problems:

There are no doors because they suck and don't work.

When I did a final test the exit hole was appearing at the spawn, and there is no reason for that to be happening.

The Corn Enemies in the Ice Room need to be replaced with snipers if we get them.